apps/ProjectManagement/requirements.txt
google-genai>=0.6.0
python-dotenv>=1.0.1
requests>=2.32.3
pywebview>=4.4.1
pyinstaller>=6.11.0
aiohttp>=3.13.2
openpyxl>=3.1.5
python-docx>=1.1.2
pywin32>=306
PyMuPDF>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
pillow-heif>=0.18.0
